- Added in auth.dart to abstract sign-in, sign-up and and user checking - Modified createUser to work locally since I don't have server running - Moved lock selection screen from main into approriate file - Refactored and removed username from createUserRequest, email serves as that now - Added rootpage to allow sign-in's
