+++ Mobile/lib/login_page.dart
@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
+import 'package:superbikelock/auth.dart';
 import 'primary_button.dart';
 
 import 'HTTPHelper.dart';
 
 
 class LoginPage extends StatefulWidget {
-  LoginPage({Key key, this.title, this.onSignIn}) : super(key: key);
+  LoginPage({Key key, this.title, this.onSignIn, BaseAuth auth}) : super(key: key);
 
   final String title;
 //  final BaseAuth auth;
@@ -46,8 +47,8 @@ class _LoginPageState extends State<LoginPage> {
         HTTPHelper createUserRequest = HTTPHelper.empty();
         //TODO update with sign in funcionality instead of creating account everytime
         String userId = (_formType == FormType.login
-            ? await createUserRequest.createUser(_username, _password, _email)
-            : await createUserRequest.createUser(_username, _password, _email)) as String;
+            ? await createUserRequest.createUser(_email, _password, )
+            : await createUserRequest.createUser(_email, _password, )) as String;
         setState(() {
           _authHint = 'Signed In\n\nUser id: $userId';
         });
