+++ Mobile/lib/root_page.dart
@@ -1,0 +1,56 @@
+import 'package:flutter/material.dart';
+import 'auth.dart';
+import 'lock_selection_screen.dart';
+import 'login_page.dart';
+
+
+class RootPage extends StatefulWidget {
+  RootPage({Key key, this.auth}) : super(key: key);
+  final BaseAuth auth;
+
+  @override
+  State<StatefulWidget> createState() => new _RootPageState();
+}
+
+enum AuthStatus {
+  notSignedIn,
+  signedIn,
+}
+
+class _RootPageState extends State<RootPage> {
+
+  AuthStatus authStatus = AuthStatus.notSignedIn;
+
+  initState() {
+    super.initState();
+    widget.auth.currentUser().then((userId) {
+      setState(() {
+        authStatus = userId != null ? AuthStatus.signedIn : AuthStatus.notSignedIn;
+      });
+    });
+  }
+
+  void _updateAuthStatus(AuthStatus status) {
+    setState(() {
+      authStatus = status;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    switch (authStatus) {
+      case AuthStatus.notSignedIn:
+        return new LoginPage(
+          title: 'Flutter Login',
+          auth: widget.auth,
+          onSignIn: () => _updateAuthStatus(AuthStatus.signedIn),
+        );
+      case AuthStatus.signedIn:
+        return new MyHomePage(
+            title: "stringy",
+            auth: widget.auth,
+            onSignOut: () => _updateAuthStatus(AuthStatus.notSignedIn)
+        );
+    }
+  }
+}
